HV: treewide: fix violations of coding guideline C-TY-02
The coding guideline rule C-TY-02 requires that 'the operands of bit
operations shall be unsigned'. This patch adds explicit casts or literal
suffixes to make explicit the type of values involved in bit operations.
Explicit casts to widen integers before shift operations are also
introduced to make explicit that the variables are expanded BEFORE it is
shifted (which is already so in C99 but implicitly).

This patch has no semantic changes.

Tracked-On: #6776
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/acpi_parser/dmar_parse.c
@@ -29,8 +29,8 @@ static uint8_t get_secondary_bus(uint8_t bus, uint8_t dev, uint8_t func)
 {
 	uint32_t data;
 
-	pio_write32(PCI_CFG_ENABLE | (bus << 16U) | (dev << 11U) |
-		(func << 8U) | 0x18U, PCI_CONFIG_ADDR);
+	pio_write32(PCI_CFG_ENABLE | ((uint32_t)bus << 16U) | ((uint32_t)dev << 11U) |
+		((uint32_t)func << 8U) | 0x18U, PCI_CONFIG_ADDR);
 
 	data = pio_read32(PCI_CONFIG_DATA);
 
@@ -134,8 +134,8 @@ static int32_t handle_one_drhd(struct acpi_dmar_hardware_unit *acpi_drhd, struct
 
 		/* Disable GPU IOMMU due to gvt-d hasn’t been enabled on APL yet. */
 		if (is_apl_platform()) {
-			if (((drhd->segment << 16U) |
-		     	     (dev_scope->bus << 8U) |
+			if ((((uint32_t)drhd->segment << 16U) |
+		     	     ((uint32_t)dev_scope->bus << 8U) |
 		     	     dev_scope->devfun) == CONFIG_GPU_SBDF) {
 				drhd->ignore = true;
 			}
